Use tmpdir not $key.* in CWD
Previously this did `rm $key.*`, which made me nervous. I know it is
correct now, but what about future-me who makes changes to the script?
What if future-me is a gosh-darn fool?

Now it creates a directory in /tmp called $tmpdir. And removes it with
`rm -r "$tmpdir"` when the script exits.

Why the double-quotes? That's sheer raging paranoia, imagining
unlikely scenarious, like tmpdir="foo ~ /",

This also fixes the problem of latex2sixel balking when the current
working directory is not writable.