Add color support, \usepackage{},  fix dvipng bug
COLORS:

One can now use \mathcolor to colorize equations. For text, use
\color{colorname} to change the foreground. Any colorname from dvips,
X11, or SVG is accepted.

USEPACKAGE:

I wanted a demo of all 68 dvips colors --- see samples/dvipscolors.ltx
--- but doing so required having a multicolumn environment. I didn't
want to add too many packages into the template, so instead I made it
easy for people to add packages as needed.

For files, just place '\usepackage' at the beginning of a line, as
usual. From the command line, any argument which starts with
'\usepackage' will be automatically moved to the latex preamble. If
one wants, a '--preamble' option is also available, but it adds no
benefit.

   latex2sixel '\usepackage{foo}'  <mathematics>
   latex2sixel --preamble '\usepackage{foo}'  <mathematics>

DVIPNG BUG AND OUTPUT ALIGNMENT

One of the biggest changes was to the output alignment, but it was
unintentional. The problem was that dvipng was occasionally generating
invalid PNG files when output was long and had many colors. Taking a
tip from the manpage, I found that \usepackage[active]{preview} fixes
the bug. However doing so also causes dvipng to ignore the offset and
bounding box (it is always -T tight)!

I put the left indentation back in by using a feature of sixel that
the image starts at the text cursor. Of course, now the units are
specified in number of character cells of the terminal instead of
centimeters.

I tried -bd for the top margin, but dvipng punts on drawing any border
when the "preview" or "geometry" packages are loaded.

Using preview does have one side-benefit: it seems to have fixed a
bug where very long graphics (like the aforementioned color table)
would display only the last "page".